feat: Working editing on uploaded large meshes
To find barycentric coordinates for arbritrary meshes, it cannot be
indexed. So we unindex the mesh when you upload it. However, since it's
now not indexed, we cannot just select a vertex index and update it. So
for now, I select an array of indices. The once resting in the same
  position, and move them all together. Kinda funny, but it works.